requirements.txt
# Core dependencies
numpy>=1.22.0
torch>=2.0.0
torchvision>=0.15.0
scipy>=1.8.0
matplotlib>=3.5.0
pillow>=9.0.0
tqdm>=4.62.0

# Simulation environment
gymnasium>=0.28.1  # Reinforcement learning environments
mujoco>=2.3.0  # MuJoCo physics engine (Apple Silicon optimized)

# Web interface - FastAPI stack
fastapi>=0.104.0
uvicorn>=0.23.0
websockets>=11.0.0
python-socketio>=5.8.0
python-multipart>=0.0.6
jinja2>=3.1.2
aiofiles>=23.1.0
starlette>=0.26.0  # FastAPI dependency
pydantic>=2.0.0   # Data validation for API
itsdangerous>=2.0.0  # For session/cookie handling

# 3D visualization
pywavefront>=1.3.0  # For loading 3D models

# V-JEPA 2 specific
transformers>=4.30.0
huggingface_hub>=0.16.0  # For downloading model weights
timm>=0.6.0
einops>=0.6.0
torch>=2.0.0      # Deep learning framework
numpy>=1.23.0     # Numerical operations
pillow>=9.0.0     # Image processing
scipy>=1.10.0     # Scientific computing

# Benchmarking and analysis tools
pandas>=1.5.0     # Data analysis and manipulation
scikit-learn>=1.0.0  # Machine learning utilities for dimensionality reduction
seaborn>=0.12.0   # Statistical data visualization

# Model optimization and deployment
onnx>=1.13.0      # Open Neural Network Exchange format
onnxruntime>=1.13.0  # ONNX Runtime for efficient inference
onnxoptimizer>=0.3.10  # For optimizing ONNX models

# Language integration
tokenizers>=0.13.0  # Fast tokenizers for language models
